+++ src/core/CardTemplate.gd
@@ -184,9 +184,9 @@ func _ready() -> void:
 	# warning-ignore:return_value_discarded
 	$TargetLine/ArrowHead/Area2D.connect("area_exited", self, "_on_ArrowHead_area_exited")
 	# warning-ignore:return_value_discarded
-	$Control.connect("mouse_entered", self, "_on_Card_mouse_entered")
+	#$Control.connect("mouse_entered", self, "_on_Card_mouse_entered")
 	# warning-ignore:return_value_discarded
-	$Control.connect("mouse_exited", self, "_on_Card_mouse_exited")
+	#$Control.connect("mouse_exited", self, "_on_Card_mouse_exited")
 	# warning-ignore:return_value_discarded
 	$Control.connect("gui_input", self, "_on_Card_gui_input")
 	# warning-ignore:return_value_discarded
@@ -249,37 +249,58 @@ func _process(delta) -> void:
 		_draw_targeting_arrow()
 	_process_card_state()
 	# Having to do all these checks due to godotengine/godot#16854
-	if _is_drawer_open and not _is_drawer_hovered() and not _is_card_hovered():
-		_on_Card_mouse_exited()
+	if cfc.debug and not get_parent() in cfc.piles:
+		var stateslist = [
+			"IN_HAND",
+			"FOCUSED_IN_HAND",
+			"MOVING_TO_CONTAINER",
+			"REORGANIZING",
+			"PUSHED_ASIDE",
+			"DRAGGED",
+			"DROPPING_TO_BOARD",
+			"ON_PLAY_BOARD",
+			"FOCUSED_ON_BOARD",
+			"DROPPING_INTO_PILE",
+			"IN_PILE",
+			"IN_POPUP",
+			"FOCUSED_IN_POPUP",
+			"VIEWPORT_FOCUS",
+		]
+		$Debug.visible = true
+		$Debug/id.text = "ID:  " + str(self)
+		$Debug/state.text = "STATE: " + stateslist[state]
+		$Debug/index.text = "INDEX: " + str(get_index())
+		$Debug/parent.text = "PARENT: " + str(get_parent().name)
 
 
 
+# Triggers the focus-in effect on the card
 func _on_Card_mouse_entered() -> void:
 	# This triggers the focus-in effect on the card
-	#print(state,":enter:",get_index()) # Debug
+	#print(state,":enter:",get_index(), ":", _are_buttons_hovered()) # Debug
 	#print($Control/Tokens/Drawer/VBoxContainer.rect_size) # debug
 	# We use this variable to check if mouse thinks it changed nodes because
 	# it just entered a child node
-	if not (_are_buttons_hovered() and not _is_drawer_hovered()) or \
-			((_are_buttons_hovered() ) and not get_focus()):
-		match state:
-			IN_HAND, REORGANIZING, PUSHED_ASIDE:
-				if not cfc.card_drag_ongoing:
-					#print("focusing:",get_my_card_index()) # debug
-					interruptTweening()
-					state = FOCUSED_IN_HAND
-			ON_PLAY_BOARD:
-				state = FOCUSED_ON_BOARD
-			# The only way to mouse over a card in a pile, is when
-			# it's in a the grid popup
-			IN_POPUP:
-				state = FOCUSED_IN_POPUP
+	#if not _is_drawer_hovered():
+	match state:
+		IN_HAND, REORGANIZING, PUSHED_ASIDE:
+			if not cfc.card_drag_ongoing:
+				#print("focusing:",get_my_card_index()) # debug
+				interruptTweening()
+				state = FOCUSED_IN_HAND
+		ON_PLAY_BOARD:
+			state = FOCUSED_ON_BOARD
+		# The only way to mouse over a card in a pile, is when
+		# it's in a the grid popup
+		IN_POPUP:
+			state = FOCUSED_IN_POPUP
 
 
 func _input(event) -> void:
 	if event is InputEventMouseButton and \
 			event.get_button_index() == 1 and \
-			not event.is_pressed():
+			not event.is_pressed() \
+			and not _are_buttons_hovered():
 		if _is_targetting:
 			complete_targeting()
 		# On depressed left mouse click anywhere on the board
@@ -294,9 +315,18 @@ func _input(event) -> void:
 # A signal for whenever the player clicks on a card
 func _on_Card_gui_input(event) -> void:
 	if event is InputEventMouseButton:
+		# because of https://github.com/godotengine/godot/issues/44138
+		# we need to double check that the card which is receiving the
+		# gui input, is actually the one with the highest index.
+		# We use our mouse pointer which is tracking this info.
+		if cfc.NMAP.board.mouse_pointer.current_focused_card \
+				and self != cfc.NMAP.board.mouse_pointer.current_focused_card:
+			cfc.NMAP.board.mouse_pointer.current_focused_card._on_Card_gui_input(event)
 		# If the player left clicks, we need to see if it's a double-click
 		# or a long click
-		if event.get_button_index() == 1:
+		elif event.is_pressed() \
+				and event.get_button_index() == 1 \
+				and not _are_buttons_hovered():
 			# If it's a double-click, then it's not a card drag
 			# But rather it's script execution
 			if event.doubleclick:
@@ -304,7 +334,7 @@ func _on_Card_gui_input(event) -> void:
 				execute_scripts()
 			# If it's a long click it might be because
 			# they want to drag the card
-			elif event.is_pressed():
+			else:
 				if (cfc.focus_style != cfc.FocusStyle.VIEWPORT and
 						(state == FOCUSED_IN_HAND
 						or state == FOCUSED_ON_BOARD)) or cfc.focus_style:
@@ -324,7 +354,7 @@ func _on_Card_gui_input(event) -> void:
 						_start_dragging()
 			# If the mouse button was released we drop the dragged card
 			# This also means a card clicked once won't try to immediately drag
-		if not event.is_pressed() and event.get_button_index() == 1:
+		elif not event.is_pressed() and event.get_button_index() == 1:
 			Input.set_mouse_mode(Input.MOUSE_MODE_VISIBLE)
 			$Control.set_default_cursor_shape(Input.CURSOR_ARROW)
 			cfc.card_drag_ongoing = null
@@ -345,9 +375,9 @@ func _on_Card_gui_input(event) -> void:
 					move_to(destination)
 					_focus_completed = false
 					#emit_signal("card_dropped",self)
-		if event.is_pressed() and event.get_button_index() == 2:
+		elif event.is_pressed() and event.get_button_index() == 2:
 			initiate_targeting()
-		if not event.is_pressed() and event.get_button_index() == 2:
+		elif not event.is_pressed() and event.get_button_index() == 2:
 			complete_targeting()
 
 
@@ -361,23 +391,23 @@ func _on_Card_mouse_exited() -> void:
 	#print(_are_buttons_hovered()) # debug
 	# We use this variable to check if mouse thinks it changed nodes
 	# because it just entered a child node
-	if not _are_buttons_hovered() and not _is_drawer_hovered():
-		match state:
-			FOCUSED_IN_HAND:
-				#_focus_completed = false
-				# To avoid errors during fast player actions
-				if get_parent() in cfc.hands:
-					for c in get_parent().get_all_cards():
-						# We need to make sure afterwards all card will return
-						# to their expected positions
-						# Therefore we simply stop all tweens and reorganize the whole hand
-						c.interruptTweening()
-						# This will also set the state to IN_HAND afterwards
-						c.reorganizeSelf()
-			FOCUSED_ON_BOARD:
-				state = ON_PLAY_BOARD
-			FOCUSED_IN_POPUP:
-				state = IN_POPUP
+	#if not _is_drawer_hovered():
+	match state:
+		FOCUSED_IN_HAND:
+			#_focus_completed = false
+			# To avoid errors during fast player actions
+			if get_parent() in cfc.hands:
+				for c in get_parent().get_all_cards():
+					# We need to make sure afterwards all card will return
+					# to their expected positions
+					# Therefore we simply stop all tweens and reorganize the whole hand
+					c.interruptTweening()
+					# This will also set the state to IN_HAND afterwards
+					c.reorganizeSelf()
+		FOCUSED_ON_BOARD:
+			state = ON_PLAY_BOARD
+		FOCUSED_IN_POPUP:
+			state = IN_POPUP
 
 
 # Resizes Token Drawer to min size whenever a token is removed completely.
@@ -439,8 +469,9 @@ func _on_View_pressed() -> void:
 # Triggers when a card hovers over another card while being dragged
 #
 # It takes care to highlight potential cards which can serve as hosts
-func _on_Card_area_entered(card: Card) -> void:
-	if (card and is_attachment and not card in attachments):
+func _on_Card_area_entered(area: Area2D) -> void:
+	if (area as Card and is_attachment and not area in attachments):
+		var card : Card = area
 		# The below 'if', checks if the card we're hovering has already been considered
 		# If it hasn't we add it to an array of all the cards we're hovering
 		# onto at the moment, at the same time
@@ -452,17 +483,19 @@ func _on_Card_area_entered(card: Card) -> void:
 			# We add the card to the list which tracks out simultaneous hovers
 			_potential_cards.append(card)
 			# We sort all potential cards by their index
-			_potential_cards.sort_custom(CardIndexSorter,"sort_index_ascending")
+			_potential_cards.sort_custom(CardFrameworkUtils,"sort_index_ascending")
 			# Finally we use a method which  handles changing highlights on the
 			# top index card
 			highlight_potential_card(cfc.HOST_HOVER_COLOUR)
 
 
+
 # Triggers when a card stops hovering over another
 #
 # It clears potential highlights and adjusts potential cards as hosts
-func _on_Card_area_exited(card: Card) -> void:
-	if card:
+func _on_Card_area_exited(area: Area2D) -> void:
+	if area as Card:
+		var card : Card = area
 		# We only do any step if the card we exited was already considered and
 		# was on the board
 		# And only while this is the card being dragged
@@ -484,7 +517,7 @@ func _on_Card_area_exited(card: Card) -> void:
 func _on_ArrowHead_area_entered(card: Card) -> void:
 	if card and not card in _potential_cards:
 		_potential_cards.append(card)
-		_potential_cards.sort_custom(CardIndexSorter,"sort_index_ascending")
+		_potential_cards.sort_custom(CardFrameworkUtils,"sort_index_ascending")
 		highlight_potential_card(cfc.TARGET_HOVER_COLOUR)
 
 
@@ -717,6 +750,17 @@ func set_card_rotation(value: int, toggle := false, start_tween := true, check :
 	# and a toggle was not requested, we consider we did nothing
 	elif value == card_rotation and not toggle:
 		retcode = _ReturnCode.OK
+		# We add this check because hand oval rotation
+		# does not change the card_rotation property
+		# so we ensure that the displayed rotation of a card
+		# which is not in hand, matches our expectations
+		if get_parent() != cfc.NMAP.hand \
+				and cfc.hand_use_oval_shape \
+				and $Control.rect_rotation != 0.0 \
+				and not $Tween.is_active():
+			_add_tween_rotation($Control.rect_rotation,value)
+			if start_tween:
+				$Tween.start()
 	else:
 		# If the toggle was specified then if the card matches the requested
 		# rotation, we reset it to 0 degrees
@@ -725,6 +769,14 @@ func set_card_rotation(value: int, toggle := false, start_tween := true, check :
 
 		# We modify the card only if this is not a cost dry-run
 		if not check:
+			card_rotation = value
+			# If the value is 0 but the card is in an oval hand, we ensure the actual
+			# rotation we apply to the card will be their hand oval rotation
+			if value == 0 \
+					and get_parent() == cfc.NMAP.hand \
+					and cfc.hand_use_oval_shape:
+				value = int(_recalculate_rotation())
+			$CollisionShape2D.rotation_degrees = value
 			# We make sure to remove other tweens of the same type
 			# to avoid a deadlock
 			# There's no way to rotate the Area2D node,
@@ -735,7 +787,6 @@ func set_card_rotation(value: int, toggle := false, start_tween := true, check :
 			# available tween, instead of immediately.
 			if start_tween:
 				$Tween.start()
-			card_rotation = value
 			#$Control/Tokens.rotation_degrees = -value # need to figure this out
 			# When the card actually changes orientation
 			# We report that it changed.
@@ -796,6 +847,8 @@ func move_to(targetHost: Node2D,
 			# The end position is always the final position the card would be
 			# inside the hand
 			_target_position = recalculate_position()
+			card_rotation = 0
+			_target_rotation = _recalculate_rotation()
 			state = MOVING_TO_CONTAINER
 			emit_signal("card_moved_to_hand",
 					self,
@@ -828,6 +881,7 @@ func move_to(targetHost: Node2D,
 				# of the stack position the card would have
 				# (this means how far up in the pile the card would appear)
 				_target_position = targetHost.get_stack_position(self)
+				_target_rotation = 0.0
 				state = MOVING_TO_CONTAINER
 				emit_signal("card_moved_to_pile",
 						self,
@@ -923,7 +977,6 @@ func execute_scripts(
 		signal_details: Dictionary = {}):
 	var card_scripts
 	var sceng = null
-
 	# If scripts have been defined directly in this Card object
 	# They take precedence over CardScriptDefinitions.gd
 	#
@@ -947,16 +1000,15 @@ func execute_scripts(
 	# We select which scripts to run from the card, based on it state
 	var state_exec := "NONE"
 	match state:
-		ON_PLAY_BOARD,FOCUSED_ON_BOARD:
+		ON_PLAY_BOARD, FOCUSED_ON_BOARD:
 			# We assume only faceup cards can execute scripts on the board
 			if is_faceup:
 				state_exec ="board"
-		IN_HAND,FOCUSED_IN_HAND:
+		IN_HAND, FOCUSED_IN_HAND, PUSHED_ASIDE:
 				state_exec ="hand"
-		IN_POPUP,FOCUSED_IN_POPUP:
+		IN_POPUP, FOCUSED_IN_POPUP:
 				state_exec ="pile"
 	state_scripts = card_scripts.get(state_exec, [])
-
 	# Here we check for confirmation of optional trigger effects
 	# There should be an SP.KEY_IS_OPTIONAL definition per state
 	# E.g. if board scripts are optional, but hand scripts are not
@@ -987,7 +1039,6 @@ func execute_scripts(
 		else: state_scripts = []
 		# Garbage cleanup
 		choices_menu.queue_free()
-
 
 	# To avoid unnecessary operations
 	# we evoce the ScriptingEngine only if we have something to execute
@@ -1080,6 +1131,7 @@ func reorganizeSelf() ->void:
 	match state:
 		IN_HAND, FOCUSED_IN_HAND, PUSHED_ASIDE:
 			_target_position = recalculate_position()
+			_target_rotation = _recalculate_rotation()
 			state = REORGANIZING
 	# This second match is  to prevent from changing the state
 	# when we're doing fancy movement
@@ -1090,6 +1142,7 @@ func reorganizeSelf() ->void:
 	match state:
 		MOVING_TO_CONTAINER:
 			_target_position = recalculate_position()
+			_target_rotation = _recalculate_rotation()
 
 
 # Stops existing card animations then makes sure they're
@@ -1294,18 +1347,34 @@ func complete_targeting() -> void:
 # * When set to false, card cannot receive inputs anymore
 #    (this is useful when card is in motion or in a pile)
 # * When set to true, card can receive inputs again
-func set_mouse_filters(value = true) -> void:
+func set_control_mouse_filters(value = true) -> void:
 	var control_filter := 0
-	var all_filter := 1
 	if not value:
 		control_filter = 2
-		all_filter = 2
 	# We do a comparison first, to make sure we avoid unnecessary operations
 	if $Control.mouse_filter != control_filter:
 		$Control.mouse_filter = control_filter
-		for n in $Control/ManipulationButtons.get_children():
-			if n as Button:
-				n.mouse_filter = all_filter
+	if monitorable != value:
+		monitorable = value
+
+
+# Changes the hosted Manipulation button node mouse filters
+#
+# * When set to false, buttons cannot receive inputs anymore
+#    (this is useful when card is in hand or a pile)
+# * When set to true, buttons can receive inputs again
+func set_manipulation_button_mouse_filters(value = true) -> void:
+	var button_filter := 1
+	# We also want the buttons disabled while the card is being targeted
+	# with a tarteting arrow.
+	# We do not want to activate the buttons when a player is trying to
+	# select the card for an effect.
+	if not value or $Control/FocusHighlight.modulate == cfc.TARGET_HOVER_COLOUR:
+		button_filter = 2
+	# We do a comparison first, to make sure we avoid unnecessary operations
+	for button in $Control/ManipulationButtons.get_children():
+		if button as Button and button.mouse_filter != button_filter:
+			button.mouse_filter = button_filter
 
 
 # Get card position in hand by index
@@ -1314,14 +1383,6 @@ func recalculate_position(index_diff = null) -> Vector2:
 	if cfc.hand_use_oval_shape:
 		return _recalculate_position_use_oval(index_diff)
 	return _recalculate_position_use_rectangle(index_diff)
-
-
-class CardIndexSorter:
-	# Used with sort_custom to find the highest child index among multiple cards
-	static func sort_index_ascending(c1: Card, c2: Card):
-		if c1.get_my_card_index() < c2.get_my_card_index():
-			return true
-		return false
 
 
 # Makes attachments always move with their parent around the board
@@ -1347,34 +1408,11 @@ func _organize_attachments() -> void:
 			# We don't want to try and move it if it's still tweening.
 			# But if it isn't, we make sure it always follows its parent
 			if not card.get_node('Tween').is_active() and \
-					card.state == ON_PLAY_BOARD:
+					card.state in [ON_PLAY_BOARD,FOCUSED_ON_BOARD]:
 				card.global_position = global_position + \
 						Vector2(0,(attach_index + 1) \
 						* $Control.rect_size.y \
 						* cfc.ATTACHMENT_OFFSET)
-
-
-# We're using this helper function, to allow our mouse-position relevant code
-# to work during integration testing
-# Returns either the adjusted global mouse position
-# or a fake mouse position provided by integration testing
-func _determine_global_mouse_pos() -> Vector2:
-	var mouse_position
-	var zoom = Vector2(1,1)
-	# We have to do the below offset hack due to godotengine/godot#30215
-	# This is caused because we're using a viewport node and
-	# scaling the game in full-creen.
-	if get_viewport().has_node("Camera2D"):
-		zoom = get_viewport().get_node("Camera2D").zoom
-	var offset_mouse_position = \
-			get_tree().current_scene.get_global_mouse_position() \
-			- get_viewport_transform().origin
-	offset_mouse_position *= zoom
-	#var scaling_offset = get_tree().get_root().get_node('Main').get_viewport().get_size_override() * OS.window_size
-	if cfc.UT: mouse_position = cfc.NMAP.board._UT_mouse_position
-	else: mouse_position = offset_mouse_position
-	return mouse_position
-
 
 
 # Returns the global mouse position but ensures it does not exit the
@@ -1382,7 +1420,7 @@ func _determine_global_mouse_pos() -> Vector2:
 #
 # Returns the adjusted global_mouse_position
 func _determine_board_position_from_mouse() -> Vector2:
-	var targetpos: Vector2 = _determine_global_mouse_pos()
+	var targetpos: Vector2 = cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos()
 	if targetpos.x + $Control.rect_size.x * scale.x >= get_viewport().size.x:
 		targetpos.x = get_viewport().size.x - $Control.rect_size.x * scale.x
 	if targetpos.x < 0:
@@ -1413,9 +1451,10 @@ func _determine_target_position_from_mouse() -> void:
 
 
 # Instructs the card to move aside for another card enterring focus
-func _pushAside(targetpos: Vector2) -> void:
+func _pushAside(targetpos: Vector2, target_rotation: float) -> void:
 	interruptTweening()
 	_target_position = targetpos
+	_target_rotation = target_rotation
 	state = PUSHED_ASIDE
 
 # Pick up a card to drag around with the mouse.
@@ -1428,12 +1467,18 @@ func _start_dragging() -> void:
 	# in full-creen.
 	if not cfc.UT:
 		if ProjectSettings.get("display/window/stretch/mode") != 'disabled':
-			get_tree().current_scene.get_viewport().warp_mouse(global_position)
+			get_tree().current_scene.get_viewport().warp_mouse(global_position + Vector2(5,5))
 		# However the above messes things if we don't have stretch mode,
 		# so we ignore it then
 		else:
 			get_viewport().warp_mouse(global_position)
 	state = DRAGGED
+	# We check if the card was already overlapping with other card
+	# before we started dragging. If so, we activate the code
+	# which checks for potential hosts, on all these cards to make sure
+	# we don't miss any.
+	for obj in get_overlapping_areas():
+		_on_Card_area_entered(obj)
 	if get_parent() in cfc.hands:
 		# While we're dragging the card from hand, we want the other cards
 		# to move to their expected position in hand
@@ -1502,20 +1547,25 @@ func _clear_attachment_status() -> void:
 # This is all necessary as a workaround for
 # https://github.com/godotengine/godot/issues/16854
 #
-# Returns true if the mouse is hovering over the buttons, else false
+# Returns true if the mouse is hovering any of the buttons, else false
 func _are_buttons_hovered() -> bool:
 	var ret = false
-	if (_determine_global_mouse_pos().x
-			>= $Control/ManipulationButtons.rect_global_position.x and
-			_determine_global_mouse_pos().y
-			>= $Control/ManipulationButtons.rect_global_position.y and
-			_determine_global_mouse_pos().x
-			<= $Control/ManipulationButtons.rect_global_position.x
-			+ $Control/ManipulationButtons.rect_size.x and
-			_determine_global_mouse_pos().y
-			<= $Control/ManipulationButtons.rect_global_position.y
-			+ $Control/ManipulationButtons.rect_size.y):
-		ret = true
+	for button in $Control/ManipulationButtons.get_children():
+		if button as Button \
+				and button.is_hovered() \
+				and button.modulate.a == 1:
+			ret = true
+#	if (cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().x
+#			>= $Control/ManipulationButtons.rect_global_position.x and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().y
+#			>= $Control/ManipulationButtons.rect_global_position.y and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().x
+#			<= $Control/ManipulationButtons.rect_global_position.x
+#			+ $Control/ManipulationButtons.rect_size.x and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().y
+#			<= $Control/ManipulationButtons.rect_global_position.y
+#			+ $Control/ManipulationButtons.rect_size.y):
+#		ret = true
 	return(ret)
 
 
@@ -1526,19 +1576,7 @@ func _are_buttons_hovered() -> bool:
 #
 # Returns true if the mouse is hovering over the token drawer, else false
 func _is_drawer_hovered() -> bool:
-	var ret = false
-	if (_determine_global_mouse_pos().x
-		>= $Control/Tokens/Drawer.rect_global_position.x and
-		_determine_global_mouse_pos().y
-		>= $Control/Tokens/Drawer.rect_global_position.y and
-		_determine_global_mouse_pos().x
-		<= $Control/Tokens/Drawer.rect_global_position.x
-		+ $Control/Tokens/Drawer.rect_size.x and
-		_determine_global_mouse_pos().y
-		<= $Control/Tokens/Drawer.rect_global_position.y
-		+ $Control/Tokens/Drawer.rect_size.y):
-		ret = true
-	return(ret)
+	return($Control/Tokens._is_hovered())
 
 
 # Detects when the mouse is still hovering over the card
@@ -1549,16 +1587,17 @@ func _is_drawer_hovered() -> bool:
 # Returns true if the mouse is hovering over the card, else false
 func _is_card_hovered() -> bool:
 	var ret = false
-	if (_determine_global_mouse_pos().x
-			>= $Control.rect_global_position.x and
-			_determine_global_mouse_pos().y
-			>= $Control.rect_global_position.y and
-			_determine_global_mouse_pos().x
-			<= $Control.rect_global_position.x
-			+ $Control.rect_size.x and
-			_determine_global_mouse_pos().y
-			<= $Control.rect_global_position.y
-			+ $Control.rect_size.y):
+#	if (cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().x
+#			>= $Control.rect_global_position.x and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().y
+#			>= $Control.rect_global_position.y and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().x
+#			<= $Control.rect_global_position.x
+#			+ $Control.rect_size.x and
+#			cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos().y
+#			<= $Control.rect_global_position.y
+#			+ $Control.rect_size.y):
+	if cfc.NMAP.board.mouse_pointer in get_overlapping_areas():
 		ret = true
 	#print(ret)
 	return(ret)
@@ -1629,7 +1668,7 @@ func _draw_targeting_arrow() -> void:
 	$TargetLine.clear_points()
 	# The final position is the mouse position,
 	# but we offset it by the position of the card center on the map
-	var final_point =  _determine_global_mouse_pos() - (position + $Control.rect_size/2)
+	var final_point =  cfc.NMAP.board.mouse_pointer.determine_global_mouse_pos() - (position + $Control.rect_size/2)
 	var curve = Curve2D.new()
 #		var middle_point = centerpos + (get_global_mouse_position() - centerpos)/2
 #		var middle_dir_to_vpcenter = middle_point.direction_to(get_viewport().size/2)
@@ -1754,16 +1793,18 @@ func _process_card_state() -> void:
 		IN_HAND:
 			z_index = 0
 			set_focus(false)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
+			set_card_rotation(0)
 			# warning-ignore:return_value_discarded
+			# When we have an oval shape, we ensure the cards stay
+			# in the rotation expected of their position
 			if cfc.hand_use_oval_shape:
-				# if not $Tween.is_active():
-				# 	_target_rotation  = _recalculate_rotation()
-				# 	_add_tween_rotation($Control.rect_rotation,_target_rotation)
-				# 	$Tween.start()
-				set_card_rotation(0)
-			else:
-				set_card_rotation(0)
+				_target_rotation  = _recalculate_rotation()
+				if not $Tween.is_active() \
+						and $Control.rect_rotation != _target_rotation:
+					_add_tween_rotation($Control.rect_rotation,_target_rotation)
+					$Tween.start()
 
 		FOCUSED_IN_HAND:
 			# Used when card is focused on by the mouse hovering over it.
@@ -1771,7 +1812,8 @@ func _process_card_state() -> void:
 			# always over its neighbours
 			z_index = 1
 			set_focus(true)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			set_card_rotation(0,false,false)
 			if not $Tween.is_active() and \
@@ -1793,7 +1835,11 @@ func _process_card_state() -> void:
 						# how close neighbours they are.
 						# Closest neighbours (1 card away) are pushed more
 						# than further neighbours.
-						neighbour_card._pushAside(neighbour_card.recalculate_position(neighbour_index_diff))
+						neighbour_card._pushAside(
+								neighbour_card.recalculate_position(
+									neighbour_index_diff),
+								neighbour_card._recalculate_rotation(
+									neighbour_index_diff))
 						neighbours.append(neighbour_card)
 				for c in get_parent().get_all_cards():
 					if not c in neighbours and c != self:
@@ -1826,7 +1872,8 @@ func _process_card_state() -> void:
 			# (i.e. deck to hand, hand to discard etc)
 			z_index = 0
 			set_focus(false)
-			set_mouse_filters(false)
+			set_control_mouse_filters(false)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			# set_card_rotation(0,false,false)
 			if not $Tween.is_active():
@@ -1887,7 +1934,6 @@ func _process_card_state() -> void:
 				# We need to check again, just in case it's been reorganized instead.
 				if state == MOVING_TO_CONTAINER:
 					_add_tween_position(position, _target_position, 0.35,Tween.TRANS_SINE, Tween.EASE_IN_OUT)
-					_target_rotation  = _recalculate_rotation()
 					_add_tween_rotation($Control.rect_rotation,_target_rotation)
 					$Tween.start()
 					yield($Tween, "tween_all_completed")
@@ -1898,7 +1944,8 @@ func _process_card_state() -> void:
 			# Used when reorganizing the cards in the hand
 			z_index = 0
 			set_focus(false)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			set_card_rotation(0,false,false)
 			if not $Tween.is_active():
@@ -1906,7 +1953,6 @@ func _process_card_state() -> void:
 				_add_tween_position(position, _target_position, 0.4)
 				if not scale.is_equal_approx(Vector2(1,1)):
 					_add_tween_scale(scale, Vector2(1,1),0.4)
-				_target_rotation  = _recalculate_rotation()
 				_add_tween_rotation($Control.rect_rotation,_target_rotation)
 				$Tween.start()
 				state = IN_HAND
@@ -1915,13 +1961,14 @@ func _process_card_state() -> void:
 			# Used when card is being pushed aside due to the focusing of a neighbour.
 			z_index = 0
 			set_focus(false)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
-			set_card_rotation(0,false,false)
 			if not $Tween.is_active() and \
 					not position.is_equal_approx(_target_position):
 				_add_tween_position(position, _target_position, 0.3,
 						Tween.TRANS_QUART, Tween.EASE_IN)
+				_add_tween_rotation($Control.rect_rotation, _target_rotation, 0.3)
 				if not scale.is_equal_approx(Vector2(1,1)):
 					_add_tween_scale(scale, Vector2(1,1), 0.3,Tween.TRANS_QUART, Tween.EASE_IN)
 				$Tween.start()
@@ -1930,7 +1977,9 @@ func _process_card_state() -> void:
 
 		DRAGGED:
 			# Used when the card is dragged around the game with the mouse
-			set_mouse_filters(true)
+			set_focus(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			if (not $Tween.is_active() and
 				not scale.is_equal_approx(cfc.card_scale_while_dragging) and
 				get_parent() != cfc.NMAP.board):
@@ -1941,11 +1990,12 @@ func _process_card_state() -> void:
 			# because if the player drags the cursor outside the window and unclicks
 			# The control will not receive the mouse input
 			# and this will stay dragging forever
-			if not cfc.UT: Input.set_mouse_mode(Input.MOUSE_MODE_CONFINED)
+			if not cfc.UT:
+				Input.set_mouse_mode(Input.MOUSE_MODE_CONFINED)
 			$Control.set_default_cursor_shape(Input.CURSOR_CROSS)
 			# We set the card to be centered on the mouse cursor to allow
 			# the player to properly understand where it will go once dropped.
-			global_position = _determine_board_position_from_mouse()# - $Control.rect_size/2 * scale
+			global_position = _determine_board_position_from_mouse() - Vector2(5,5)
 			_organize_attachments()
 			# We want to keep the token drawer closed during movement
 			if _is_drawer_open:
@@ -1954,7 +2004,8 @@ func _process_card_state() -> void:
 		ON_PLAY_BOARD:
 			# Used when the card is idle on the board
 			set_focus(false)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(true)
 			if not $Tween.is_active() and \
 					not scale.is_equal_approx(cfc.PLAY_AREA_SCALE):
 				_add_tween_scale(scale, cfc.PLAY_AREA_SCALE, 0.3,
@@ -1973,7 +2024,8 @@ func _process_card_state() -> void:
 			_organize_attachments()
 
 		DROPPING_TO_BOARD:
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			# Used when dropping the cards to the table
 			# When dragging the card, the card is slightly behind the mouse cursor
 			# so we tween it to the right location
@@ -2003,7 +2055,8 @@ func _process_card_state() -> void:
 			# Used when card is focused on by the mouse hovering over it while it is on the board.
 			# The below tween shows the container manipulation buttons when you hover over them
 			set_focus(true)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(true)
 			if not _buttons_tween.is_active() and \
 					$Control/ManipulationButtons.modulate[3] != 1:
 				_buttons_tween.remove_all()
@@ -2016,7 +2069,7 @@ func _process_card_state() -> void:
 #		DROPPING_INTO_PILE:
 #			# Used when dropping the cards into a container (Deck, Discard etc)
 #			set_focus(false)
-#			set_mouse_filters(false)
+#			set_control_mouse_filters(false)
 #			if not $Tween.is_active():
 #				var intermediate_position: Vector2
 #				if cfc.fancy_movement:
@@ -2047,7 +2100,8 @@ func _process_card_state() -> void:
 
 		IN_PILE:
 			set_focus(false)
-			set_mouse_filters(false)
+			set_control_mouse_filters(false)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			set_card_rotation(0)
 			if scale != Vector2(1,1):
@@ -2057,7 +2111,8 @@ func _process_card_state() -> void:
 			# We make sure that a card in a popup stays in its position
 			# Unless moved
 			set_focus(false)
-			set_mouse_filters(true)
+			set_control_mouse_filters(true)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			set_card_rotation(0)
 			if modulate[3] != 1:
@@ -2075,7 +2130,8 @@ func _process_card_state() -> void:
 
 		VIEWPORT_FOCUS:
 			set_focus(false)
-			set_mouse_filters(false)
+			set_control_mouse_filters(false)
+			set_manipulation_button_mouse_filters(false)
 			# warning-ignore:return_value_discarded
 			set_card_rotation(0)
 			$Control/ManipulationButtons.modulate[3] = 0
@@ -2163,10 +2219,13 @@ func _get_angle_by_index(index_diff = null):
 	half = (hand_size - 1) / 2.0
 	var card_angle_max: float = 15
 	var card_angle_min: float = 5
+	# Angle between cards
 	var card_angle = max(min(60/hand_size,card_angle_max),card_angle_min)
+	# When foucs hand, the card needs to be offset by a certain angle
+	# The current practice is just to find a suitable expression function, if there is a better function, please replace this function: - sign(index_diff) * (1.95-0.3*index_diff*index_diff) * min(card_angle,5)
 	if index_diff != null:
 		return 90 + (half - index) * card_angle \
-				- sign(index_diff) * (- index_diff * index_diff + 10)
+				- sign(index_diff) * (1.95-0.3*index_diff*index_diff) * min(card_angle,5)
 	else:
 		return 90 + (half - index) * card_angle
 
@@ -2179,6 +2238,7 @@ func _get_oval_angle_by_index(
 		hor_rad = null,
 		ver_rad = null):
 	if not angle:
+		# Get the angle from the point on the oval to the center of the oval
 		angle = _get_angle_by_index(index_diff)
 	var parent_control
 	if not hor_rad:
@@ -2191,6 +2251,7 @@ func _get_oval_angle_by_index(
 	if angle == 90:
 		card_angle = 90
 	else:
+		# Convert oval angle to normal angle
 		card_angle= rad2deg(atan(- ver_rad / hor_rad / tan(deg2rad(angle))))
 		card_angle = card_angle + 90
 	return card_angle
@@ -2201,18 +2262,35 @@ func _recalculate_position_use_oval(index_diff = null)-> Vector2:
 	var card_position_x: float = 0.0
 	var card_position_y: float = 0.0
 	var parent_control = get_parent().get_node('Control')
-	var hor_rad: float = parent_control.rect_size.x * 0.5 * 1.7
+	# Oval hor rad, rect_size.x*0.5*1.5 it’s an empirical formula, 
+	# that's been tested to feel good.
+	var hor_rad: float = parent_control.rect_size.x * 0.5 * 1.5
+	# Oval ver rad, rect_size.y * 1.5 it’s an empirical formula,
+	# that's been tested to feel good.
 	var ver_rad: float = parent_control.rect_size.y * 1.5
+	# Get the angle from the point on the oval to the center of the oval
 	var angle = _get_angle_by_index(index_diff)
 	var rad_angle = deg2rad(angle)
+	# Get the direction vector of a point on the oval
 	var oval_angle_vector = Vector2(hor_rad * cos(rad_angle),
 			- ver_rad * sin(rad_angle))
+	# Take the center point of the card as the starting point, the coordinates 
+	# of the top left corner of the card
 	var left_top = Vector2(- $Control.rect_size.x/2, - $Control.rect_size.y/2)
+	# Place the top center of the card on the oval point
 	var center_top = Vector2(0, - $Control.rect_size.y/2)
+	# Get the angle of the card, which is different from the oval angle, 
+	# the card angle is the normal angle of a certain point
 	var card_angle = _get_oval_angle_by_index(angle, null, hor_rad,ver_rad)
+	# Displacement offset due to card rotation
 	var delta_vector = left_top - center_top.rotated(deg2rad(90 - card_angle))
-	var center_x = parent_control.rect_size.x / 2 + parent_control.rect_position.x
-	var center_y = parent_control.rect_size.y * 1.5 + parent_control.rect_position.y
+	# Oval center x
+	var center_x = parent_control.rect_size.x / 2 \
+			+ parent_control.rect_position.x
+	# Oval center y, - parent_control.rect_size.y * 0.25:This method ensures that the card is moved to the proper position
+	var center_y = parent_control.rect_size.y * 1.5 \
+			+ parent_control.rect_position.y \
+			- parent_control.rect_size.y * 0.25
 	card_position_x = (oval_angle_vector.x + center_x)
 	card_position_y = (oval_angle_vector.y + center_y)
 	return(Vector2(card_position_x, card_position_y) + delta_vector)