+++ tests/UTcommon.gd
@@ -6,6 +6,7 @@ const BOARD_SCENE = preload("res://src/custom/Board.tscn")
 const MOUSE_SPEED := {
 	"fast": [10,0.3],
 	"slow": [3,0.6],
+	"debug": [1,2],
 }
 
 
@@ -70,17 +71,21 @@ func unclick_card_anywhere(card: Card) -> void:
 func drag_card(card: Card, target_position: Vector2, interpolation_speed := "fast") -> void:
 	var mouse_speed = MOUSE_SPEED[interpolation_speed][0]
 	var mouse_yield_wait = MOUSE_SPEED[interpolation_speed][1]
-	card._on_Card_mouse_entered()
+	board._UT_interpolate_mouse_move(card.global_position + Vector2(20,20),board._UT_mouse_position,mouse_speed)
+	yield(yield_for(mouse_yield_wait), YIELD)
 	click_card(card)
-	yield(yield_for(0.3), YIELD) # Wait to allow dragging to start
-	board._UT_interpolate_mouse_move(target_position,card.global_position,mouse_speed)
+	if interpolation_speed == "debug":
+		yield(yield_for(4), YIELD) # Allow for review
+	else:
+		yield(yield_for(0.3), YIELD) # Wait to allow dragging to start
+	board._UT_interpolate_mouse_move(target_position,board._UT_mouse_position,mouse_speed)
 	yield(yield_for(mouse_yield_wait), YIELD)
 
 
 func drop_card(card: Card, drop_location: Vector2) -> void:
 	var fc:= fake_click(false, drop_location)
 	card._on_Card_gui_input(fc)
-	yield(yield_to(card.get_node('Tween'), "tween_all_completed", 1), YIELD)
+	yield(yield_to(card._tween, "tween_all_completed", 1), YIELD)
 
 
 # Takes care of simple drag&drop requests
@@ -100,7 +105,10 @@ func target_card(source, target, interpolation_speed := "fast") -> void:
 		yield(yield_for(0.6), YIELD)
 	# We need to offset a bit towards the card rect, to ensure the arrow
 	# Area2D collides
-	board._UT_interpolate_mouse_move(target.global_position + Vector2(10,10),
+	var extra_offset = Vector2(10,10)
+	if target.card_rotation in [90,270]:
+		extra_offset = Vector2(10,100)
+	board._UT_interpolate_mouse_move(target.global_position + extra_offset,
 			source.global_position,mouse_speed)
 	yield(yield_for(mouse_yield_wait), YIELD)
 	unclick_card_anywhere(source)
@@ -112,4 +120,8 @@ func table_move(card: Card, pos: Vector2) -> void:
 	if cfc.fancy_movement:
 		yield(yield_to(card._tween, "tween_all_completed", 0.5), YIELD)
 
-
+func move_mouse(target_position: Vector2, interpolation_speed := "fast") -> void:
+	var mouse_speed = MOUSE_SPEED[interpolation_speed][0]
+	var mouse_yield_wait = MOUSE_SPEED[interpolation_speed][1]
+	board._UT_interpolate_mouse_move(target_position,board._UT_mouse_position,mouse_speed)
+	yield(yield_for(mouse_yield_wait), YIELD)