+++ tests/integration/test_board_use.gd
@@ -8,7 +8,11 @@ func before_each():
 	yield(yield_for(1), YIELD)
 
 
-func test_card_table_drop_location_and_rotation():
+func test_card_table_drop_location_and_rotation_use_rectangle():
+	cfc.hand_use_oval_shape = false
+	for c in cfc.NMAP.hand.get_all_cards():
+		c.reorganizeSelf()
+	yield(yield_for(0.5), YIELD) # Wait to allow dragging to start		
 	# Reminder that card should not have trigger script definitions, to avoid
 	# messing with the tests
 	var card = cards[1]
@@ -58,11 +62,25 @@ func test_card_table_drop_location_and_rotation():
 	drop_card(card,board._UT_mouse_position)
 	yield(yield_to(card._tween, "tween_all_completed", 0.5), YIELD)
 	yield(yield_to(card._tween, "tween_all_completed", 0.5), YIELD)
-	if cfc.hand_use_oval_shape:
-		pass
-	else:
-		assert_eq(0.0,card.get_node("Control").rect_rotation,
-				"Rotation reset to 0 when card moved off board")
+	assert_eq(0.0,card.get_node("Control").rect_rotation,
+			"Rotation reset to 0 while card is moving to hand")
+	cfc.hand_use_oval_shape = true
+
+func test_card_table_drop_location_use_oval():
+	cfc.hand_use_oval_shape = true
+	# Reminder that card should not have trigger script definitions, to avoid
+	# messing with the tests
+	var card = cards[1]
+	yield(table_move(card, Vector2(100,200)), "completed")
+	yield(yield_to(card._tween, "tween_all_completed", 1), YIELD)
+	card.card_rotation = 180
+	yield(drag_drop(card, Vector2(100,600)), 'completed')
+	yield(yield_to(card._tween, "tween_all_completed", 0.5), YIELD)
+	yield(yield_to(card._tween, "tween_all_completed", 0.5), YIELD)
+	assert_almost_eq(12.461,card.get_node("Control").rect_rotation,2.0,
+			"Rotation reset to a hand angle when card moved off board")
+	cfc.hand_use_oval_shape = true
+
 
 func test_card_hand_drop_recovery():
 	var card = cards[1]
@@ -85,19 +103,19 @@ func test_card_drag_block_by_board_borders():
 	yield(yield_for(0.5), YIELD) # Wait to allow dragging to start
 	board._UT_interpolate_mouse_move(Vector2(-100,100),card.global_position)
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(0, 100),card.global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(-5, 95),card.global_position,Vector2(2,2),
 			"Dragged outside left viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(1300,300))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(1220, 300),card.global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(1215, 295),card.global_position,Vector2(2,2),
 			"Dragged outside right viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(800,-100))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(800, 0),card.global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(795, -5),card.global_position,Vector2(2,2),
 			"Dragged outside top viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(500,800))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(500, 624),card.global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(495, 619),card.global_position,Vector2(2,2),
 			"Dragged outside bottom viewport borders stays inside viewport")
 
 
@@ -116,6 +134,7 @@ func test_fast_card_table_drop():
 	yield(yield_for(0.6), YIELD)
 	assert_almost_eq(Vector2(1000, 300),cards[0].global_position,Vector2(2,2),
 			"Card not dragged with mouse after dropping on table")
+
 
 func test_board_to_board_move():
 	var card: Card