+++ changelog.md
@@ -2,12 +2,19 @@
 
 ## 0.12
 
+### New Features
+
 * Scipting Engine target finding now supports tutoring cards from deck
 * Scipting Engine target finding now supports affecting multiple cards from the board by filtering their properties
 * Scripting Engine now supports cards with multiple choices for abilities.
 * Scripting Engine now supports cards with optional abilities
 * Hand shape can now be oval (default) - Contribution by [@vmjcv](https://github.com/vmjcv).
 * Font-size will decrease when it amount of chars is too large, in order to stay within the defined label rectangle - Contribution by [@vmjcv](https://github.com/vmjcv).
+
+### Bugfixes
+
+* Fixed manipulation buttonsd messing with player's actions
+
 
 ## 0.11